+++ src/program/ui/InputEditorView.cpp
@@ -294,6 +294,8 @@ void InputEditorView::mousePressEvent(QMouseEvent *event)
     selectionModel()->clear();
     mouseSection = index.column();
     mouseRow = index.row();
+    mouseMinRow = mouseRow;
+    mouseMaxRow = mouseRow;
 
     /* Rewind when clicking for column */
     if (mouseSection == 0) {
@@ -335,28 +337,43 @@ void InputEditorView::mouseMoveEvent(QMouseEvent *event)
 
     event->accept();
 
-    /* Disable toggle together with rewind, because it can cause multiple
-     * rewinds because of the scrolling */
-    if (static_cast<unsigned int>(index.row()) < context->framecount)
+    /* Check if we need to toggle this input */
+    if ((index.row() >= mouseMinRow) && (index.row() <= mouseMaxRow))
         return;
 
-    /* Prevent toggle past the first toggle frame when rewinding */
+    /* Prevent toggle past the first toggle frame when rewinding, and
+     * if we started to toggle future inputs, don't trigger a rewind */
     if (index.row() < minToggleRow) {
         return;
     }
 
     int newMouseValue = mouseValue;
 
-    /* Check if we need to alternate the input state */
-    if (event->modifiers() & Qt::ControlModifier) {
-        if ((index.row() - mouseRow) % 2)
-            newMouseValue = !newMouseValue;
+    /* Toggle all cells from mouse to minRow-1, or maxRow+1 to mouse */
+    int minLoop, maxLoop;
+    if (index.row() < mouseMinRow) {
+        minLoop = index.row();
+        maxLoop = mouseMinRow - 1;
+        mouseMinRow = index.row();
+    }
+    else {
+        minLoop = mouseMaxRow + 1;
+        maxLoop = index.row();
+        mouseMaxRow = index.row();        
     }
 
-    /* Toggle the cell with the same row as the cell under the mouse */
-    QModelIndex toggle_index = inputEditorModel->index(index.row(), mouseSection);
-
-    inputEditorModel->setData(toggle_index, QVariant(newMouseValue), Qt::EditRole);
+    for (int i = minLoop; i <= maxLoop; i++) {
+        /* Check if we need to alternate the input state */
+        if (event->modifiers() & Qt::ControlModifier) {
+            if ((i - mouseRow) % 2)
+            newMouseValue = !newMouseValue;
+        }
+        
+        /* Toggle the cell with the same row as the cell under the mouse */
+        QModelIndex toggle_index = inputEditorModel->index(i, mouseSection);
+        
+        inputEditorModel->setData(toggle_index, QVariant(newMouseValue), Qt::EditRole);        
+    }
 }
 
 void InputEditorView::keyPressEvent(QKeyEvent *event)